refactor:make the component names consistent
make the component names consistent across the application

Closes #1432
