+++ src/components/common/customComponents/DeleteResource.tsx
