+++ src/components/sandbox/Execution.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import styled from 'styled-components';
-import PolicyComponent from './components/PolicyComponent';
-import ResourcesComponent from './components/ResourcesComponent';
 import { EquinorIcon } from '../common/StyledComponents';
 import DatasetConfirmed from './components/DatasetsConfirmed';
 import { SandboxObj } from '../common/interfaces';
+import Policy from './components/Policy';
+import Resources from './components/Resources';
 
 const Wrapper = styled.div`
     display: grid;
@@ -48,10 +48,10 @@ const Execution: React.FC<ExecutionProps> = ({ resources, sandbox }) => {
                     <DatasetConfirmed sandbox={sandbox} />
                 </div>
                 <div>
-                    <PolicyComponent sandbox={sandbox} />
+                    <Policy sandbox={sandbox} />
                 </div>
                 <div>
-                    <ResourcesComponent resources={resources} permissions={sandbox.permissions} />
+                    <Resources resources={resources} permissions={sandbox.permissions} />
                 </div>
             </ResourceWrapper>
         </Wrapper>
