+++ src/components/sandbox/components/ResourceItem.tsx
