+++ src/components/studyDetails/Participant.tsx
